// src/ui/slider/Slider.d.ts
import { ReactNode } from "react";
type Props = {
    duration?: number;
    height?: string;
    setCurrentIndex?: (currentIndex: number) => void;
    slides: ReactNode[];
    stylesSlide?: string;
    stylesSlider?: string;
};
export declare const Slider: ({ duration, height, setCurrentIndex, slides, stylesSlide, stylesSlider, }: Props) => import("react/jsx-runtime").JSX.Element;
export {};
//# sourceMappingURL=Slider.d.ts.map